Day 18

+++ chapter6/src/chapter6_result/MyMath.java
@@ -4,5 +4,5 @@ public class MyMath {
 	long add(long a, long b) { return a + b; }
 	long subtract(long a, long b) { return a - b; }
 	long multiply(long a, long b) { return a * b; }
-	long divide(double a, double b) { return (long)a / (long)b; }
+	double divide(double a, double b) { return a / b; }
 }
